fivek-cnn/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.0.0
rawpy>=0.18.0
tqdm>=4.64.0
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.1.0
streamlit>=1.28.0
